Add a CLI framework.
This adds a new component, UberLoggerCLI, which is added to the
UberAppConsole prefab.  If this is enabled, another TextArea is added
to the console window, for the user to enter commands.

Command handlers can be registered with UberLoggerCLI.RegisterHandler,
and they will be called through SendMessage("UberLoggerCLI_HandleCommand")
with the tokenized command.